requirements.txt
certifi==2021.5.30
commentjson==0.9.0
lark-parser==0.7.8
numpy==1.21.2
Pillow==8.3.2
pybind11==2.7.1
PyEXR==0.3.10
scipy==1.7.1
tqdm==4.62.2
